Finish extracting creation of intor page data.
